Fix (predictable) integer overflow on Solr uptime check

+++ modules/solr/src/main/scala/eu/ehri/project/search/solr/SolrSearchEngine.scala
@@ -51,7 +51,7 @@ case class SolrSearchEngine @Inject()(
     val params = Seq("action" -> "STATUS", "wt" -> "json", "core" -> "portal")
 
     ws.url(url).withQueryString(params: _*).get().map { r =>
-      (r.json \ "status" \ "portal" \ "uptime").validate[Int]
+      (r.json \ "status" \ "portal" \ "uptime").validate[Long]
         .fold(err => throw BadJson(err), _ => "ok")
     }.recover {
       case e => throw SearchEngineOffline(e.getMessage, e)
